pyproject.toml
[project]
name = "an-bot"
version = "0.1.0"
description = "Twitter(X) bot aimed to publish french AN"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "tweepy (>=4.15.0,<5.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "requests (>=2.32.3,<3.0.0)",
    "cattrs (>=24.1.3,<25.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
autoflake = "^2.3.1"
isort = "^6.0.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
